Add scanner.Service
